+++ src/contents.rs
@@ -46,7 +46,7 @@ use std::ops::{Deref, DerefMut};
 /// [`ContentsRef`]: struct.ContentsRef.html
 /// [`Deref`]: #impl-Deref-for-Contents
 /// [`DerefMut`]: #impl-DerefMut-for-Contents
-#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
+#[derive(Debug, Clone, Eq, Hash)]
 pub struct Contents {
     buffer: Buffer,
 }
@@ -216,7 +216,7 @@ impl Contents {
     /// let bytes: &[u8] = &[1, 2, 3];
     /// let contents = Contents::from_bytes(bytes);
     ///
-    /// assert_eq!(&contents as &[u8], bytes);
+    /// assert_eq!(contents.as_bytes(), bytes);
     /// ```
     pub fn from_bytes(bytes: &[u8]) -> Self {
         Self {
@@ -356,30 +356,30 @@ impl Contents {
 
 impl AsRef<[u8]> for Contents {
     fn as_ref(&self) -> &[u8] {
+        self.as_bytes()
+    }
+}
+
+impl AsRef<ContentsRef> for Contents {
+    fn as_ref(&self) -> &ContentsRef {
         self
     }
 }
 
 impl AsMut<[u8]> for Contents {
     fn as_mut(&mut self) -> &mut [u8] {
-        self
+        self.as_mut_bytes()
     }
 }
 
-impl Borrow<[u8]> for Contents {
-    fn borrow(&self) -> &[u8] {
+impl AsMut<ContentsRef> for Contents {
+    fn as_mut(&mut self) -> &mut ContentsRef {
         self
     }
 }
 
 impl Borrow<ContentsRef> for Contents {
     fn borrow(&self) -> &ContentsRef {
-        self
-    }
-}
-
-impl BorrowMut<[u8]> for Contents {
-    fn borrow_mut(&mut self) -> &mut [u8] {
         self
     }
 }
@@ -404,9 +404,12 @@ impl DerefMut for Contents {
     }
 }
 
-impl PartialEq<ContentsRef> for Contents {
-    fn eq(&self, other: &ContentsRef) -> bool {
-        self == other
+impl<T> PartialEq<T> for Contents
+where
+    T: Borrow<T>,
+{
+    fn eq(&self, other: &T) -> bool {
+        self == other.borrow()
     }
 }
 
@@ -420,7 +423,7 @@ mod tests {
             let contents = Contents::from_integer(i);
             let expected: &[u8] = &[i as u8];
 
-            assert_eq!(&contents as &[u8], expected);
+            assert_eq!(contents.as_bytes(), expected);
         }
     }
 
@@ -429,13 +432,13 @@ mod tests {
         for i in 0..0x80 {
             let contents = Contents::from_integer(i as u8);
             let expected: &[u8] = &[i as u8];
-            assert_eq!(&contents as &[u8], expected);
+            assert_eq!(contents.as_bytes(), expected);
         }
 
         for i in 0x80..=u8::MAX {
             let contents = Contents::from_integer(i as u8);
             let expected: &[u8] = &[0x00, i];
-            assert_eq!(&contents as &[u8], expected);
+            assert_eq!(contents.as_bytes(), expected);
         }
     }
 
@@ -450,7 +453,7 @@ mod tests {
             let f = i.unsigned_shr(8) as u8;
             let s = i as u8;
             let expected = &[f, s];
-            assert_eq!(&contents as &[u8], expected);
+            assert_eq!(contents.as_bytes(), expected);
         }
     }
 
@@ -463,7 +466,7 @@ mod tests {
             let s = i as u8;
             let expected: &[u8] = &[0, f, s];
 
-            assert_eq!(&contents as &[u8], expected);
+            assert_eq!(contents.as_bytes(), expected);
         }
     }
 
@@ -476,7 +479,7 @@ mod tests {
             let mut expected: [u8; 16] = [0x00; 16];
             expected[0] = 0x80;
 
-            assert_eq!(&contents as &[u8], expected);
+            assert_eq!(contents.as_bytes(), expected);
         }
 
         // i128::MAX
@@ -486,7 +489,7 @@ mod tests {
             let mut expected: [u8; 16] = [0xff; 16];
             expected[0] = 0x7f;
 
-            assert_eq!(&contents as &[u8], expected);
+            assert_eq!(contents.as_bytes(), expected);
         }
     }
 
@@ -499,7 +502,7 @@ mod tests {
             let mut expected: [u8; 17] = [0x00; 17];
             expected[1] = 0x80;
 
-            assert_eq!(&contents as &[u8], expected);
+            assert_eq!(contents.as_bytes(), expected);
         }
 
         // u128::MAX
@@ -509,7 +512,7 @@ mod tests {
             let mut expected: [u8; 17] = [0xff; 17];
             expected[0] = 0x00;
 
-            assert_eq!(&contents as &[u8], expected);
+            assert_eq!(contents.as_bytes(), expected);
         }
     }
 
@@ -857,13 +860,13 @@ mod tests {
         // True
         {
             let contents = ContentsRef::from_bool(true);
-            assert_eq!(&[0xff], contents as &[u8]);
+            assert_eq!(&[0xff], contents.as_bytes());
         }
 
         // false
         {
             let contents = ContentsRef::from_bool(false);
-            assert_eq!(&[0x00], contents as &[u8]);
+            assert_eq!(&[0x00], contents.as_bytes());
         }
     }
 
@@ -944,13 +947,13 @@ mod tests {
         // True
         {
             let contents = Contents::from_bool(true);
-            assert_eq!(&[0xff], &contents as &[u8]);
+            assert_eq!(&[0xff], contents.as_bytes());
         }
 
         // false
         {
             let contents = Contents::from_bool(false);
-            assert_eq!(&[0x00], &contents as &[u8]);
+            assert_eq!(&[0x00], contents.as_bytes());
         }
     }
 }